refactor(config): use native uint16 for database timeout

+++ config.h
@@ -11,7 +11,7 @@ extern int bwt_ttl;
 extern char *bwt_key;
 
 extern char *database_file;
-extern int database_timeout;
+extern uint16_t database_timeout;
 
 extern int log_level;
 extern int log_requests;
